test(bruno-js): consolidate the sandbox specs around the shared harness
Inline the spec timing values at their use sites; move the OOM fixtures
behind a runOomScript helper and rejection capture behind
collectUnhandledRejections; pin that the sync eval runs on the trapped
module and that the failed replacement build really started
